Remove tutorial for catboost integration
The tutorial described how gradient boosting (a machine learning
technique) can be run from ClickHouse using Yandex's catboost library.
Because the catboost integration code was recently rewritten, the
tutorial is now largely obsolete. The steps required to setup the
catboost integration were moved to the documentation of the
"catboostEvalaluate()" function in the official SQL function reference.